Split nb and can traits to separate crates.
