+++ embedded-can/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "embedded-can"
+version = "0.4.0"
+edition = "2018"
+
+description = "HAL traits for Controller Area Network (CAN) devices."
+categories = ["embedded", "hardware-support", "no-std"]
+documentation = "https://docs.rs/embedded-can"
+keywords = ["hal", "IO"]
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/rust-embedded/embedded-hal"
+
+[dependencies]
+embedded-hal = { version = "=1.0.0-alpha.8", path = ".." }
+nb = "1"
