+++ embedded-hal-async/src/i2c.rs
@@ -17,7 +17,7 @@
 //! `SevenBitAddress` has been set as default mode and thus can be omitted if desired.
 
 use core::future::Future;
-pub use embedded_hal::i2c::blocking::Operation;
+pub use embedded_hal::i2c::Operation;
 pub use embedded_hal::i2c::{
     AddressMode, Error, ErrorKind, ErrorType, NoAcknowledgeSource, SevenBitAddress, TenBitAddress,
 };
