+++ src/i2c.rs
@@ -22,7 +22,7 @@
 //! Here is an example of an embedded-hal implementation of the `Write` trait
 //! for both modes:
 //! ```
-//! # use embedded_hal::i2c::{ErrorKind, ErrorType, SevenBitAddress, TenBitAddress, blocking::{I2c, Operation}};
+//! # use embedded_hal::i2c::{ErrorKind, ErrorType, SevenBitAddress, TenBitAddress, I2c, Operation};
 //! /// I2C0 hardware peripheral which supports both 7-bit and 10-bit addressing.
 //! pub struct I2c0;
 //!
@@ -97,7 +97,7 @@
 //! For demonstration purposes the address mode parameter has been omitted in this example.
 //!
 //! ```
-//! # use embedded_hal::i2c::{blocking::I2c, Error};
+//! # use embedded_hal::i2c::{I2c, Error};
 //! const ADDR: u8  = 0x15;
 //! # const TEMP_REGISTER: u8 = 0x1;
 //! pub struct TemperatureSensorDriver<I2C> {
@@ -120,7 +120,7 @@
 //! ### Device driver compatible only with 10-bit addresses
 //!
 //! ```
-//! # use embedded_hal::i2c::{Error, TenBitAddress, blocking::I2c};
+//! # use embedded_hal::i2c::{Error, TenBitAddress, I2c};
 //! const ADDR: u16  = 0x158;
 //! # const TEMP_REGISTER: u8 = 0x1;
 //! pub struct TemperatureSensorDriver<I2C> {
@@ -256,202 +256,196 @@ impl AddressMode for SevenBitAddress {}
 
 impl AddressMode for TenBitAddress {}
 
-/// Blocking I2C traits
-pub mod blocking {
+/// Transactional I2C operation.
+///
+/// Several operations can be combined as part of a transaction.
+#[derive(Debug, PartialEq)]
+pub enum Operation<'a> {
+    /// Read data into the provided buffer
+    Read(&'a mut [u8]),
+    /// Write data from the provided buffer
+    Write(&'a [u8]),
+}
 
-    use super::{AddressMode, ErrorType, SevenBitAddress};
-
-    /// Transactional I2C operation.
+/// Blocking I2C
+pub trait I2c<A: AddressMode = SevenBitAddress>: ErrorType {
+    /// Reads enough bytes from slave with `address` to fill `buffer`
     ///
-    /// Several operations can be combined as part of a transaction.
-    #[derive(Debug, PartialEq)]
-    pub enum Operation<'a> {
-        /// Read data into the provided buffer
-        Read(&'a mut [u8]),
-        /// Write data from the provided buffer
-        Write(&'a [u8]),
+    /// # I2C Events (contract)
+    ///
+    /// ``` text
+    /// Master: ST SAD+R        MAK    MAK ...    NMAK SP
+    /// Slave:           SAK B0     B1     ... BN
+    /// ```
+    ///
+    /// Where
+    ///
+    /// - `ST` = start condition
+    /// - `SAD+R` = slave address followed by bit 1 to indicate reading
+    /// - `SAK` = slave acknowledge
+    /// - `Bi` = ith byte of data
+    /// - `MAK` = master acknowledge
+    /// - `NMAK` = master no acknowledge
+    /// - `SP` = stop condition
+    fn read(&mut self, address: A, buffer: &mut [u8]) -> Result<(), Self::Error>;
+
+    /// Writes bytes to slave with address `address`
+    ///
+    /// # I2C Events (contract)
+    ///
+    /// ``` text
+    /// Master: ST SAD+W     B0     B1     ... BN     SP
+    /// Slave:           SAK    SAK    SAK ...    SAK
+    /// ```
+    ///
+    /// Where
+    ///
+    /// - `ST` = start condition
+    /// - `SAD+W` = slave address followed by bit 0 to indicate writing
+    /// - `SAK` = slave acknowledge
+    /// - `Bi` = ith byte of data
+    /// - `SP` = stop condition
+    fn write(&mut self, address: A, bytes: &[u8]) -> Result<(), Self::Error>;
+
+    /// Writes bytes to slave with address `address`
+    ///
+    /// # I2C Events (contract)
+    ///
+    /// Same as the `write` method
+    fn write_iter<B>(&mut self, address: A, bytes: B) -> Result<(), Self::Error>
+    where
+        B: IntoIterator<Item = u8>;
+
+    /// Writes bytes to slave with address `address` and then reads enough bytes to fill `buffer` *in a
+    /// single transaction*
+    ///
+    /// # I2C Events (contract)
+    ///
+    /// ``` text
+    /// Master: ST SAD+W     O0     O1     ... OM     SR SAD+R        MAK    MAK ...    NMAK SP
+    /// Slave:           SAK    SAK    SAK ...    SAK          SAK I0     I1     ... IN
+    /// ```
+    ///
+    /// Where
+    ///
+    /// - `ST` = start condition
+    /// - `SAD+W` = slave address followed by bit 0 to indicate writing
+    /// - `SAK` = slave acknowledge
+    /// - `Oi` = ith outgoing byte of data
+    /// - `SR` = repeated start condition
+    /// - `SAD+R` = slave address followed by bit 1 to indicate reading
+    /// - `Ii` = ith incoming byte of data
+    /// - `MAK` = master acknowledge
+    /// - `NMAK` = master no acknowledge
+    /// - `SP` = stop condition
+    fn write_read(
+        &mut self,
+        address: A,
+        bytes: &[u8],
+        buffer: &mut [u8],
+    ) -> Result<(), Self::Error>;
+
+    /// Writes bytes to slave with address `address` and then reads enough bytes to fill `buffer` *in a
+    /// single transaction*
+    ///
+    /// # I2C Events (contract)
+    ///
+    /// Same as the `write_read` method
+    fn write_iter_read<B>(
+        &mut self,
+        address: A,
+        bytes: B,
+        buffer: &mut [u8],
+    ) -> Result<(), Self::Error>
+    where
+        B: IntoIterator<Item = u8>;
+
+    /// Execute the provided operations on the I2C bus.
+    ///
+    /// Transaction contract:
+    /// - Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.
+    /// - Data from adjacent operations of the same type are sent after each other without an SP or SR.
+    /// - Between adjacent operations of a different type an SR and SAD+R/W is sent.
+    /// - After executing the last operation an SP is sent automatically.
+    /// - If the last operation is a `Read` the master does not send an acknowledge for the last byte.
+    ///
+    /// - `ST` = start condition
+    /// - `SAD+R/W` = slave address followed by bit 1 to indicate reading or 0 to indicate writing
+    /// - `SR` = repeated start condition
+    /// - `SP` = stop condition
+    fn transaction<'a>(
+        &mut self,
+        address: A,
+        operations: &mut [Operation<'a>],
+    ) -> Result<(), Self::Error>;
+
+    /// Execute the provided operations on the I2C bus (iterator version).
+    ///
+    /// Transaction contract:
+    /// - Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.
+    /// - Data from adjacent operations of the same type are sent after each other without an SP or SR.
+    /// - Between adjacent operations of a different type an SR and SAD+R/W is sent.
+    /// - After executing the last operation an SP is sent automatically.
+    /// - If the last operation is a `Read` the master does not send an acknowledge for the last byte.
+    ///
+    /// - `ST` = start condition
+    /// - `SAD+R/W` = slave address followed by bit 1 to indicate reading or 0 to indicate writing
+    /// - `SR` = repeated start condition
+    /// - `SP` = stop condition
+    fn transaction_iter<'a, O>(&mut self, address: A, operations: O) -> Result<(), Self::Error>
+    where
+        O: IntoIterator<Item = Operation<'a>>;
+}
+
+impl<A: AddressMode, T: I2c<A>> I2c<A> for &mut T {
+    fn read(&mut self, address: A, buffer: &mut [u8]) -> Result<(), Self::Error> {
+        T::read(self, address, buffer)
     }
 
-    /// Blocking I2C
-    pub trait I2c<A: AddressMode = SevenBitAddress>: ErrorType {
-        /// Reads enough bytes from slave with `address` to fill `buffer`
-        ///
-        /// # I2C Events (contract)
-        ///
-        /// ``` text
-        /// Master: ST SAD+R        MAK    MAK ...    NMAK SP
-        /// Slave:           SAK B0     B1     ... BN
-        /// ```
-        ///
-        /// Where
-        ///
-        /// - `ST` = start condition
-        /// - `SAD+R` = slave address followed by bit 1 to indicate reading
-        /// - `SAK` = slave acknowledge
-        /// - `Bi` = ith byte of data
-        /// - `MAK` = master acknowledge
-        /// - `NMAK` = master no acknowledge
-        /// - `SP` = stop condition
-        fn read(&mut self, address: A, buffer: &mut [u8]) -> Result<(), Self::Error>;
-
-        /// Writes bytes to slave with address `address`
-        ///
-        /// # I2C Events (contract)
-        ///
-        /// ``` text
-        /// Master: ST SAD+W     B0     B1     ... BN     SP
-        /// Slave:           SAK    SAK    SAK ...    SAK
-        /// ```
-        ///
-        /// Where
-        ///
-        /// - `ST` = start condition
-        /// - `SAD+W` = slave address followed by bit 0 to indicate writing
-        /// - `SAK` = slave acknowledge
-        /// - `Bi` = ith byte of data
-        /// - `SP` = stop condition
-        fn write(&mut self, address: A, bytes: &[u8]) -> Result<(), Self::Error>;
-
-        /// Writes bytes to slave with address `address`
-        ///
-        /// # I2C Events (contract)
-        ///
-        /// Same as the `write` method
-        fn write_iter<B>(&mut self, address: A, bytes: B) -> Result<(), Self::Error>
-        where
-            B: IntoIterator<Item = u8>;
-
-        /// Writes bytes to slave with address `address` and then reads enough bytes to fill `buffer` *in a
-        /// single transaction*
-        ///
-        /// # I2C Events (contract)
-        ///
-        /// ``` text
-        /// Master: ST SAD+W     O0     O1     ... OM     SR SAD+R        MAK    MAK ...    NMAK SP
-        /// Slave:           SAK    SAK    SAK ...    SAK          SAK I0     I1     ... IN
-        /// ```
-        ///
-        /// Where
-        ///
-        /// - `ST` = start condition
-        /// - `SAD+W` = slave address followed by bit 0 to indicate writing
-        /// - `SAK` = slave acknowledge
-        /// - `Oi` = ith outgoing byte of data
-        /// - `SR` = repeated start condition
-        /// - `SAD+R` = slave address followed by bit 1 to indicate reading
-        /// - `Ii` = ith incoming byte of data
-        /// - `MAK` = master acknowledge
-        /// - `NMAK` = master no acknowledge
-        /// - `SP` = stop condition
-        fn write_read(
-            &mut self,
-            address: A,
-            bytes: &[u8],
-            buffer: &mut [u8],
-        ) -> Result<(), Self::Error>;
-
-        /// Writes bytes to slave with address `address` and then reads enough bytes to fill `buffer` *in a
-        /// single transaction*
-        ///
-        /// # I2C Events (contract)
-        ///
-        /// Same as the `write_read` method
-        fn write_iter_read<B>(
-            &mut self,
-            address: A,
-            bytes: B,
-            buffer: &mut [u8],
-        ) -> Result<(), Self::Error>
-        where
-            B: IntoIterator<Item = u8>;
-
-        /// Execute the provided operations on the I2C bus.
-        ///
-        /// Transaction contract:
-        /// - Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.
-        /// - Data from adjacent operations of the same type are sent after each other without an SP or SR.
-        /// - Between adjacent operations of a different type an SR and SAD+R/W is sent.
-        /// - After executing the last operation an SP is sent automatically.
-        /// - If the last operation is a `Read` the master does not send an acknowledge for the last byte.
-        ///
-        /// - `ST` = start condition
-        /// - `SAD+R/W` = slave address followed by bit 1 to indicate reading or 0 to indicate writing
-        /// - `SR` = repeated start condition
-        /// - `SP` = stop condition
-        fn transaction<'a>(
-            &mut self,
-            address: A,
-            operations: &mut [Operation<'a>],
-        ) -> Result<(), Self::Error>;
-
-        /// Execute the provided operations on the I2C bus (iterator version).
-        ///
-        /// Transaction contract:
-        /// - Before executing the first operation an ST is sent automatically. This is followed by SAD+R/W as appropriate.
-        /// - Data from adjacent operations of the same type are sent after each other without an SP or SR.
-        /// - Between adjacent operations of a different type an SR and SAD+R/W is sent.
-        /// - After executing the last operation an SP is sent automatically.
-        /// - If the last operation is a `Read` the master does not send an acknowledge for the last byte.
-        ///
-        /// - `ST` = start condition
-        /// - `SAD+R/W` = slave address followed by bit 1 to indicate reading or 0 to indicate writing
-        /// - `SR` = repeated start condition
-        /// - `SP` = stop condition
-        fn transaction_iter<'a, O>(&mut self, address: A, operations: O) -> Result<(), Self::Error>
-        where
-            O: IntoIterator<Item = Operation<'a>>;
+    fn write(&mut self, address: A, bytes: &[u8]) -> Result<(), Self::Error> {
+        T::write(self, address, bytes)
     }
 
-    impl<A: AddressMode, T: I2c<A>> I2c<A> for &mut T {
-        fn read(&mut self, address: A, buffer: &mut [u8]) -> Result<(), Self::Error> {
-            T::read(self, address, buffer)
-        }
+    fn write_iter<B>(&mut self, address: A, bytes: B) -> Result<(), Self::Error>
+    where
+        B: IntoIterator<Item = u8>,
+    {
+        T::write_iter(self, address, bytes)
+    }
 
-        fn write(&mut self, address: A, bytes: &[u8]) -> Result<(), Self::Error> {
-            T::write(self, address, bytes)
-        }
+    fn write_read(
+        &mut self,
+        address: A,
+        bytes: &[u8],
+        buffer: &mut [u8],
+    ) -> Result<(), Self::Error> {
+        T::write_read(self, address, bytes, buffer)
+    }
 
-        fn write_iter<B>(&mut self, address: A, bytes: B) -> Result<(), Self::Error>
-        where
-            B: IntoIterator<Item = u8>,
-        {
-            T::write_iter(self, address, bytes)
-        }
+    fn write_iter_read<B>(
+        &mut self,
+        address: A,
+        bytes: B,
+        buffer: &mut [u8],
+    ) -> Result<(), Self::Error>
+    where
+        B: IntoIterator<Item = u8>,
+    {
+        T::write_iter_read(self, address, bytes, buffer)
+    }
 
-        fn write_read(
-            &mut self,
-            address: A,
-            bytes: &[u8],
-            buffer: &mut [u8],
-        ) -> Result<(), Self::Error> {
-            T::write_read(self, address, bytes, buffer)
-        }
+    fn transaction<'a>(
+        &mut self,
+        address: A,
+        operations: &mut [Operation<'a>],
+    ) -> Result<(), Self::Error> {
+        T::transaction(self, address, operations)
+    }
 
-        fn write_iter_read<B>(
-            &mut self,
-            address: A,
-            bytes: B,
-            buffer: &mut [u8],
-        ) -> Result<(), Self::Error>
-        where
-            B: IntoIterator<Item = u8>,
-        {
-            T::write_iter_read(self, address, bytes, buffer)
-        }
-
-        fn transaction<'a>(
-            &mut self,
-            address: A,
-            operations: &mut [Operation<'a>],
-        ) -> Result<(), Self::Error> {
-            T::transaction(self, address, operations)
-        }
-
-        fn transaction_iter<'a, O>(&mut self, address: A, operations: O) -> Result<(), Self::Error>
-        where
-            O: IntoIterator<Item = Operation<'a>>,
-        {
-            T::transaction_iter(self, address, operations)
-        }
+    fn transaction_iter<'a, O>(&mut self, address: A, operations: O) -> Result<(), Self::Error>
+    where
+        O: IntoIterator<Item = Operation<'a>>,
+    {
+        T::transaction_iter(self, address, operations)
     }
 }
